move icons to feature folders
