+++ frontend/src/modules/app/sidebar/ShutdownIcon.tsx
