+++ frontend/src/modules/app/sidebar/SidebarClickable.tsx
@@ -14,8 +14,8 @@ import { commitRequest } from '../../../store/controller/actions';
 import { getParametersIsVentilating } from '../../../store/controller/selectors/measurements';
 import { ScreenStatusRequest } from '../../../store/proto/mcu_pb';
 import { MessageType } from '../../../store/proto/types';
-import { HomeIcon } from '../../icons';
-import ShutdownIcon from '../../icons/ShutdownIcon';
+import HomeIcon from './HomeIcon';
+import ShutdownIcon from './ShutdownIcon';
 import { DASHBOARD_ROUTE, QUICKSTART_ROUTE } from '../navigation/constants';
 import Navigation from '../navigation/Navigation';
