+++ frontend/src/modules/app/sidebar/index.ts
@@ -1,2 +1,5 @@
 export * from './Sidebar';
 export * from './SidebarClickable';
+export { default as HomeIcon } from './HomeIcon';
+export { default as PufferfishIcon } from './PufferfishIcon';
+export { default as ShutdownIcon } from './ShutdownIcon';
