+++ frontend/src/modules/app/toolbar/Alarms.tsx
@@ -16,7 +16,7 @@ import {
 } from '../../../store/controller/selectors';
 import { getEventType } from '../../event-logs/EventType';
 import { LOGS_ROUTE } from '../navigation/constants';
-import { BellIcon } from '../../icons';
+import { BellIcon } from './icons';
 import EventLogModal from '../../event-logs/EventLog';
 import {
   AlarmMuteCountdown,
