+++ frontend/src/modules/app/toolbar/icons/ClockIcon.tsx
