+++ frontend/src/modules/app/toolbar/icons/Power25Icon.tsx
