+++ frontend/src/modules/app/toolbar/icons/Power50Icon.tsx
