+++ frontend/src/modules/app/toolbar/icons/Power75Icon.tsx
