+++ frontend/src/modules/app/toolbar/icons/PowerChargingIcon.tsx
