+++ frontend/src/modules/app/toolbar/icons/PowerFullIcon.tsx
