+++ frontend/src/modules/app/toolbar/icons/index.ts
@@ -1,0 +1,1 @@
+export { default as BellIcon } from './BellIcon';
