+++ frontend/src/modules/shared/icons/BackIcon.tsx
