+++ frontend/src/modules/shared/icons/BellNoneIcon.tsx
