+++ frontend/src/modules/shared/icons/ChartIcon.tsx
