+++ frontend/src/modules/shared/icons/GearIcon.tsx
