+++ frontend/src/modules/shared/icons/LungsIcon.tsx
