+++ frontend/src/modules/shared/icons/TrendDownIcon.tsx
