+++ frontend/src/modules/shared/icons/TrendUpIcon.tsx
