+++ frontend/src/modules/shared/icons/index.ts
@@ -15,12 +15,8 @@
  *
  */
 export { default as BackIcon } from './BackIcon';
-export { default as BellIcon } from './BellIcon';
 export { default as ChartIcon } from './ChartIcon';
 export { default as GearIcon } from './GearIcon';
-export { default as HomeIcon } from './HomeIcon';
 export { default as LungsIcon } from './LungsIcon';
-export { default as PufferfishIcon } from './PufferfishIcon';
-export { default as ShutdownIcon } from './ShutdownIcon';
 export { default as TrendDownIcon } from './TrendDownIcon';
 export { default as TrendUpIcon } from './TrendUpIcon';
